requirements.txt
# Pipeline & optional web backend
pyyaml
opencv-python-headless
numpy
Pillow
scikit-image
requests
fastapi
uvicorn[standard]

# Text OCR (local, default)
pytesseract

# Optional: PaddleOCR (better for 中英文混排). Set ocr.engine: "paddleocr" in config.
# Use PaddlePaddle 3.2.2 + PaddleOCR 3.x (3.3.0+ has CPU oneDNN bug, will fallback to Tesseract):
#   pip install paddlepaddle==3.2.2 paddleocr   # or paddlepaddle-gpu==3.2.2

# Optional: formula recognition (LaTeX). Install for full text pipeline.
# pip install pix2text   # and onnxruntime or onnxruntime-gpu for GPU
# Optional: background removal (RMBG). Install for IconPictureProcessor.
# pip install onnxruntime  # or onnxruntime-gpu
# PyTorch and SAM3: install separately; see README (PyTorch + SAM3 source).
